Remove and suppress warnings in genlib and tests (#252)
* Squash warnings on windows

* Crank up warning level to /W4 on MSVC

* Fix w4 warnings in MSVC

* Remove some clang warnings

* Remove psubvec - only used in one place and does not add real value.

* Shut up warning about overload hiding in paftl (we know about it and the
only way to get rid of it is getting rid of paftl)

* Suppress warnings on mac as well.

+++ genlib/containerutils.h
@@ -50,14 +50,14 @@ bool addIfNotExists(std::map<K, V> &map, const K &key, const V &value) {
 }
 
 template<typename K, typename V>
-typename std::map<K, V>::const_iterator getMapAtIndex(const std::map<K, V> &m, int idx) {
+typename std::map<K, V>::const_iterator getMapAtIndex(const std::map<K, V> &m, size_t idx) {
     auto iter = m.begin();
     std::advance(iter, idx);
     return iter;
 }
 
 template<typename K, typename V>
-typename std::map<K, V>::iterator getMapAtIndex(std::map<K, V> &m, int idx) {
+typename std::map<K, V>::iterator getMapAtIndex(std::map<K, V> &m, size_t idx) {
     auto iter = m.begin();
     std::advance(iter, idx);
     return iter;
